+++ class/GaiaAlpha/Model/Menu.php
@@ -8,9 +8,6 @@ class Menu extends BaseModel
 
     public static function findByLocation(string $location)
     {
-        $db = \GaiaAlpha\Controller\DbController::getPdo();
-        $stmt = $db->prepare("SELECT * FROM menus WHERE location = ? LIMIT 1");
-        $stmt->execute([$location]);
-        return $stmt->fetch(\PDO::FETCH_ASSOC);
+        return BaseModel::fetch("SELECT * FROM menus WHERE location = ? LIMIT 1", [$location]);
     }
 }